# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  constants.BudgetStatus:
    enum:
    - Draft
    - Active
    - Locked
    - Archived
    type: string
    x-enum-varnames:
    - BudgetStatusDraft
    - BudgetStatusActive
    - BudgetStatusLocked
    - BudgetStatusArchived
  constants.PaymentMethod:
    enum:
    - Cash
    - BankTransfer
    - Credit
    - Company
    type: string
    x-enum-varnames:
    - PaymentMethodCash
    - PaymentMethodBankTransfer
    - PaymentMethodCredit
    - PaymentMethodCompany
  constants.RoleName:
    enum:
    - Admin
    - FinanceManager
    - ProjectManager
    - Accountant
    - Owner
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleFinanceManager
    - RoleProjectManager
    - RoleAccountant
    - RoleOwner
...
  dto.AdjustBudgetRequest:
    properties:
      reason:
        type: string
      total_amount:
        type: number
    required:
    - reason
    - total_amount
    type: object
  dto.ApproveExpenseRequest:
    properties:
      approval_comment:
        maxLength: 512
        type: string
    type: object
  dto.CreateBudgetRequest:
    properties:
      name:
        maxLength: 128
        type: string
      project_id:
        type: string
      status:
        $ref: '#/definitions/constants.BudgetStatus'
      total_amount:
        type: number
    required:
    - name
    - project_id
    - total_amount
    type: object
  dto.CreateExpenseRequest:
    properties:
      amount:
        type: number
      attachment_url:
        maxLength: 512
        type: string
      budget_item_id:
        type: integer
      description:
        maxLength: 512
        type: string
      expense_date:
        type: string
      invoice_no:
        maxLength: 128
        type: string
      payment_method:
        $ref: '#/definitions/constants.PaymentMethod'
      supplier_id:
        type: integer
    required:
    - amount
    - budget_item_id
    - expense_date
    - payment_method
    type: object
...
  dto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponse:
    properties:
      expires_in:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserInfo'
    type: object
  dto.PayExpenseRequest:
    properties:
      payment_date:
        type: string
    type: object
  dto.RejectExpenseRequest:
    properties:
      approval_comment:
        maxLength: 512
        type: string
    required:
    - approval_comment
    type: object
  dto.UpdateBudgetRequest:
    properties:
      name:
        maxLength: 128
        type: string
      status:
        $ref: '#/definitions/constants.BudgetStatus'
      total_amount:
        minimum: 0
        type: number
    type: object
...
  dto.UserInfo:
    properties:
      display_name:
        type: string
      id:
        type: integer
      role:
        $ref: '#/definitions/constants.RoleName'
      username:
        type: string
    type: object
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:19306
info:
  contact:
    email: support@example.com
    name: API Support
  description: 装修预算与支出管理 API 服务，提供预算编制、支出审批、供应商管理与对账结算能力。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: Renovation Budget API
  version: 1.0.0
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用预置角色账号登录并获取 JWT。
      parameters:
      - description: 登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Body'
      summary: 登录
      tags:
      - auth
  /budgets:
    get:
      parameters:
      - description: 项目ID
        in: query
        name: project_id
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 查询预算表列表
      tags:
      - budgets
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBudgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 创建预算表
      tags:
      - budgets
  /budgets/{id}:
    delete:
      parameters:
      - description: 预算表ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 删除预算表
      tags:
      - budgets
    get:
      parameters:
      - description: 预算表ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 获取预算表详情
      tags:
      - budgets
    put:
      consumes:
      - application/json
      parameters:
      - description: 预算表ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBudgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 更新预算表
      tags:
      - budgets
  /budgets/{id}/adjust:
    post:
      consumes:
      - application/json
      parameters:
      - description: 预算表ID
        in: path
        name: id
        required: true
        type: integer
      - description: 调整请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdjustBudgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 调整预算总额
      tags:
      - budgets
...
  /expenses:
    get:
      parameters:
      - description: 审批状态
        in: query
        name: status
        type: string
      - description: 预算项ID
        in: query
        name: budget_item_id
        type: integer
      - description: 供应商ID
        in: query
        name: supplier_id
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 查询支出记录列表
      tags:
      - expenses
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 创建支出记录
      tags:
      - expenses
  /expenses/{id}:
    get:
      parameters:
      - description: 支出记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 获取支出记录详情
      tags:
      - expenses
  /expenses/{id}/approve:
    post:
      consumes:
      - application/json
      parameters:
      - description: 支出记录ID
        in: path
        name: id
        required: true
        type: integer
      - description: 审批意见
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ApproveExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 审批通过支出
      tags:
      - expenses
  /expenses/{id}/pay:
    post:
      consumes:
      - application/json
      parameters:
      - description: 支出记录ID
        in: path
        name: id
        required: true
        type: integer
      - description: 付款日期
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PayExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 确认付款
      tags:
      - expenses
  /expenses/{id}/reject:
    post:
      consumes:
      - application/json
      parameters:
      - description: 支出记录ID
        in: path
        name: id
        required: true
        type: integer
      - description: 驳回意见
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RejectExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 驳回支出
      tags:
      - expenses
  /expenses/{id}/submit:
    post:
      parameters:
      - description: 支出记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: 提交支出
      tags:
      - expenses
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"